requirements.txt
beautifulsoup4==4.12.2
colorama==0.4.6
geopy==2.4.0
gTTS==2.4.0
imutils==0.5.4
matplotlib==3.3.4
numpy==1.24.3
opencv_python==4.5.3.56
pandas==2.1.2
Pillow==10.1.0
pytesseract==0.3.10
python_telegram_bot==20.7
pytz==2021.1
requests==2.31.0
scikit_image==0.18.1
scour==0.38.2
telegram==0.0.1
